Math.Equation

Documentation

type Solution a = Vector a

Type synonym for the solution of an equation system, represented as a vector.

solve :: (Fractional a, Eq a) => Matrix a -> Vector a -> Solution a

Solve an equation system represented by a matrix and a vector.

backwards :: (Eq a, Fractional a) => Array Int (Row a) -> Row a -> Solution a

Solve the equation system by backward substitution.

check :: (Eq a, Num a) => Int -> Array Int (Row a) -> Int

Check if a solution is impossible due to a zero pivot element.