Math.Matrix

Documentation

type Row a = Array Int a

Type synonym for a row in a matrix.

data Matrix a

Data type representing a matrix.

Constructors

Matrix

Fields

rows :: Int

Number of rows in the matrix.

cols :: Int

Number of columns in the matrix.

elements :: Array (Int, Int) a

Array storing the elements of the matrix.

get :: Matrix a -> Int -> Int -> a

Lookup the value at a specific row and column.

asRows :: Matrix a -> Array Int (Row a)

Convert a matrix to an array of rows.

toList :: Matrix a -> [a]

Convert a matrix to a list.

shape :: Matrix a -> (Int, Int)

Return the shape of the matrix (number of rows and columns).

reshape :: Int -> Int -> [a] -> Matrix a

Return a matrix of the specified shape with elements from the provided list (wrapping if necessary).

innerProduct :: Num a => (a -> a -> a) -> (b -> c -> a) -> Matrix b -> Matrix c -> Matrix a

Perform the inner product of two matrices using provided addition and multiplication functions.

transpose :: Matrix a -> Matrix a

Transpose a matrix (rows become columns, and vice versa).