Math.Vector

Documentation

type Vector a = Array Int a

Type synonym for a vector represented as an array.

showV :: Show a => Vector a -> String

Convert a vector to its string representation.

toMatrix :: Vector a -> Matrix a

Convert a vector to a single-column matrix.

fromList :: [a] -> Vector a

Convert a list to a vector.

fromMatrix :: Matrix a -> Vector a

Convert a matrix to a vector.

binop :: (a -> a -> a) -> Vector a -> Vector a -> Vector a

Function for binary operators on vectors.

(+.) :: Num a => Vector a -> Vector a -> Vector a

Vector addition.

(-.) :: Num a => Vector a -> Vector a -> Vector a

Vector subtraction.

(*.) :: Num a => Vector a -> Vector a -> Vector a

Vector multiplication.

(<.>) :: Vector a -> Vector a -> Vector a

Vector concatenation.

dot :: Num a => Array Int a -> Array Int a -> a

Dot product of two vectors.

projection :: (Eq a, Fractional a) => Array Int a -> Array Int a -> Array Int a

Calculate the projection of a vector onto another vector with normalization.