Primitives.Generator

Documentation

rng :: StdGen

Fixed random number generator seed.

iota :: Int -> Matrix Int

Generate an iota matrix (a row vector with consecutive integer elements).

deal :: Int -> Int -> Matrix Int

Generate a matrix with unique random integers in the range [1, max].

primes :: Int -> Int -> Matrix Int

Generate a matrix with the first prime numbers arranged in r rows and c columns.