Primitives.VTransform

Documentation

pick :: Int -> Vector a -> a

Pick the element at the specified index in the vector.

tally :: Vector a -> Int

Get the number of elements in the vector.

vhead :: Vector a -> a

Get the first element of the vector.

behead :: Vector a -> Vector a

Remove the first element from the vector.

hat :: Vector a -> Vector a

Reverse the vector.

vtail :: Vector a -> a

Get the last element of the vector.

vnub :: Eq a => Vector a -> Vector a

Remove duplicate elements from the vector while preserving the order.

encode :: (Integral a, Show a) => a -> Vector a -> Vector String

Encode the elements of the vector using a given base and return a vector of strings.