Utilities.Latex.Styles

Documentation

data LatexDoc

Represents a LaTeX document configuration.

Constructors

LatexDoc

Fields

title :: String

Title of the document.

author :: String

Author of the document.

style :: LatexStyle

LaTeX style configuration.

data LatexStyle

Represents the style configuration for a LaTeX document.

Constructors

LatexStyle

Fields

sectionSize :: FontSize

Font size for section titles.

bodySize :: FontSize

Font size for the body text.

twoColumns :: Bool

Whether the document uses two columns layout.

vectorStyle :: TexVector

Style for vectors in the document.

matrixStyle :: TexMatrix

Style for matrices in the document.

data FontSize

Represents different font sizes available for LaTeX documents.

Constructors

Small

Medium

Large

Huge

data TexVector

Represents the style configuration for vectors in the document.

Constructors

OneSize

Later

data TexMatrix

Represents the style configuration for matrices in the document.

Constructors

OneSizeM

LaterM

minimalStyle :: LatexStyle

Predefined LaTeX style with larger section titles and medium body text.

twoColumnStyle :: LatexStyle

Predefined LaTeX style with medium section titles, medium body text, and two columns layout.

compactStyle :: LatexStyle

Predefined LaTeX style with smaller section titles, medium body text, and two columns layout.