Utilities.Logger

Documentation

data Entry

Represents an entry in the logger with a tag (message) and a value.

Constructors

Entry

Fields

tag :: String

The tag (message) of the log entry.

val :: String

The value associated with the log entry.

type Logger a = Writer [Entry] a

Type synonym for the logger using the Writer monad.

runLogger :: Logger a -> (a, [Entry])

Runs the logger and returns the result along with the list of log entries.

log :: Show a => String -> a -> Logger ()

Write a log entry with a message and a value.

nb :: String -> Logger ()

Log a note (no associated value).

calc :: String -> Logger ()

Log a calculation (no associated value).